move convesation card
