+++ src/components/ConversationCard/index.ts
@@ -1,0 +1,2 @@
+export * from './ConversationCard';
+export type { ConversationCardProps } from './ConversationCard';
