+++ src/components/ConversationsList.tsx
@@ -1,7 +1,7 @@
 import { Grid, Typography } from '@material-ui/core';
 import React from 'react';
 import { useConversations } from '../hooks/useConversations';
-import { ConversationCard } from './ConversationCard';
+import { ConversationCard } from './ConversationCard/ConversationCard';
 import PageTitle from './PageTitle';
 import Loader from './Loader';
 import { ItsBrokenIcon } from './ItsBrokenIcon';
